[feat] add more owner checks on diary get calls

+++ backend/app/diary/lib.go
@@ -1,5 +1,7 @@
 package diary
 
+// TODO: Please write some tests please
+
 import (
 	"errors"
 	"net/http"
@@ -25,15 +27,15 @@ func GetAllDiariesOfUser(c echo.Context, db *gorm.DB) error {
 	if err != nil {
 		return c.JSON(http.StatusBadRequest, echo.Map{"error": "Authentication error"})
 	}
-	diaries, err := GetAllDiariesOfUsername(db, u.Username)
+	diaries, err := UncheckedGetAllDiariesOfUsername(db, u.Username)
 	if err != nil {
 		return c.JSON(http.StatusInternalServerError, echo.Map{"error": "Database error"})
 	}
 	return c.JSON(http.StatusOK, diaries)
 }
 
-// UNSAFE. the username is not validated here.
-func GetAllDiariesOfUsername(db *gorm.DB, creator string) ([]Diary, error) {
+// UNSAFE: the username is not validated here.
+func UncheckedGetAllDiariesOfUsername(db *gorm.DB, creator string) ([]Diary, error) {
 	diaries := []Diary{}
 	if err := db.Where("creator = ?", creator).Find(&diaries).Error; err != nil {
 		return nil, errors.New("Database error: failed to get diaries of a user")
@@ -41,20 +43,37 @@ func GetAllDiariesOfUsername(db *gorm.DB, creator string) ([]Diary, error) {
 	return diaries, nil
 }
 
-func GetDiaryByID(c echo.Context, db *gorm.DB) error {
-	id, err := strconv.Atoi(c.Param("id"))
-	if err != nil {
-		return c.JSON(http.StatusBadRequest, echo.Map{"error": "Invalid diary ID"})
-	}
+func UncheckedGetDiaryByID(db *gorm.DB, id int) (*Diary, error) {
 	diary := &Diary{}
 	if err := db.First(diary, id).Error; err != nil {
-		return c.JSON(http.StatusNotFound, echo.Map{"error": "Diary not found"})
+		return nil, errors.New("not found")
+	}
+	return diary, nil
+}
+func GetDiaryByID(c echo.Context, db *gorm.DB) error {
+	u, err := user.FromEchoContext(db, c)
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, echo.Map{"error": "Authentication error"})
+	}
+	id, err := strconv.Atoi(c.Param("id"))
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, echo.Map{"error": "Invalid id format"})
+	}
+	diary, err := UncheckedGetDiaryByID(db, id)
+	if diary.CreatorID != u.ID {
+		return c.JSON(http.StatusNotFound, echo.Map{"error": "This diary is not yours"})
 	}
 	return c.JSON(http.StatusOK, diary)
 }
 
 func CreateDiary(c echo.Context, db *gorm.DB) error {
+	u, err := user.FromEchoContext(db, c)
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, echo.Map{"error": "Invalid user. consider creating new one."})
+	}
 	diary := new(Diary)
+	// UNTESTED: I'm not sure how this works or how this should be done. Test this function.
+	diary.Creator = *u
 	if err := c.Bind(diary); err != nil {
 		return c.JSON(http.StatusBadRequest, echo.Map{"error": "Failed to bind diary data"})
 	}
@@ -65,24 +84,51 @@ func CreateDiary(c echo.Context, db *gorm.DB) error {
 }
 
 func UpdateDiary(c echo.Context, db *gorm.DB) error {
-	id, err := strconv.Atoi(c.Param("id"))
+	u, err := user.FromEchoContext(db, c)
 	if err != nil {
-		return c.JSON(http.StatusBadRequest, echo.Map{"error": "Invalid diary ID"})
+		return c.JSON(http.StatusBadRequest, echo.Map{"error": "Invalid user. create new one"})
 	}
-	diary := &Diary{}
-	if err := db.First(diary, id).Error; err != nil {
-		return c.JSON(http.StatusNotFound, echo.Map{"error": "Diary not found"})
+	id, err := strconv.Atoi(c.Param("id"))
+	if err != nil || id < 0 {
+		return c.JSON(http.StatusBadRequest, echo.Map{"error": "Invalid formating of id or negative value"})
 	}
-	if err := c.Bind(diary); err != nil {
-		return c.JSON(http.StatusBadRequest, echo.Map{"error": "Failed to bind diary data"})
+	var newDiary Diary
+	if err := c.Bind(newDiary); err != nil {
+		return c.JSON(http.StatusBadRequest, echo.Map{"error": "Failde to bind diary"})
 	}
-	if err := db.Save(diary).Error; err != nil {
+	diary, err := UncheckedGetDiaryByID(db, id)
+	if err != nil {
+		return c.JSON(http.StatusNotFound, echo.Map{"error": "not found"})
+	}
+	if diary.CreatorID != u.ID {
+		return c.JSON(http.StatusUnauthorized, echo.Map{"error": "You don't own this diary"})
+	}
+	err = UncheckedUpdateDiary(db, uint(id), newDiary)
+	if err != nil {
 		return c.JSON(http.StatusInternalServerError, echo.Map{"error": "Failed to update diary"})
 	}
 	return c.JSON(http.StatusOK, diary)
 }
 
+func UncheckedUpdateDiary(db *gorm.DB, id uint, newDiary Diary) error {
+	if id != newDiary.ID {
+		return errors.New("unmatched ID")
+	}
+	diary := &Diary{}
+	if err := db.First(diary, id).Error; err != nil {
+		return err
+	}
+	if err := db.Save(newDiary).Error; err != nil {
+		return err
+	}
+	return nil
+}
+
 func DeleteDiary(c echo.Context, db *gorm.DB) error {
+	u, err := user.FromEchoContext(db, c)
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, echo.Map{"error": "Authentication error: user not found"})
+	}
 	id, err := strconv.Atoi(c.Param("id"))
 	if err != nil {
 		return c.JSON(http.StatusBadRequest, echo.Map{"error": "Invalid diary ID"})
@@ -90,6 +136,9 @@ func DeleteDiary(c echo.Context, db *gorm.DB) error {
 	diary := &Diary{}
 	if err := db.First(diary, id).Error; err != nil {
 		return c.JSON(http.StatusNotFound, echo.Map{"error": "Diary not found"})
+	}
+	if diary.CreatorID != u.ID {
+		return c.JSON(http.StatusUnauthorized, echo.Map{"error": "You don't own this diary"})
 	}
 	if err := db.Delete(diary).Error; err != nil {
 		return c.JSON(http.StatusInternalServerError, echo.Map{"error": "Failed to delete diary"})
